+++ WiseAuth.Sample/Modules/Products/ProductPermissions.cs
@@ -1,6 +1,4 @@
-using WiseAuth;
-
-namespace WiseAuth.Sample.Products;
+namespace WiseAuth.Sample.Modules.Products;
 
 [ClaimType("products")]
 public enum ProductPermissions